correcciones domingo 4

+++ src/controllers/ingresoEgreso.js
@@ -66,11 +66,11 @@ const postIngresoEgreso = async (req, res, next) => {
         egresos:
           getSaldo[getSaldo.length - 1]?.egresos + parseInt(req.body.monto),
         saldo_final:
-          getSaldo[getSaldo.length - 1]?.saldo_final === 0
-            ? getSaldo[getSaldo.length - 1]?.saldo_inicial -
-              parseInt(req.body.monto)
-            : getSaldo[getSaldo.length - 1]?.saldo_final -
-              parseInt(req.body.monto),
+        getSaldo[getSaldo.length - 1]?.saldo_final === 0
+        ? getSaldo[getSaldo.length - 1]?.saldo_inicial -
+          parseInt(req.body.monto)
+        : getSaldo[getSaldo.length - 1]?.saldo_final -
+          parseInt(req.body.monto),
       };
 
       let objIngreso = {
@@ -94,11 +94,11 @@ const postIngresoEgreso = async (req, res, next) => {
           parseInt(req.body.monto),
         egresos: "",
         saldo_final:
-          getSaldoEgreso[getSaldoEgreso.length - 1]?.saldo_final === 0
-            ? getSaldoEgreso[getSaldoEgreso.length - 1]?.saldo_inicial +
-              parseInt(req.body.monto)
-            : getSaldoEgreso[getSaldoEgreso.length - 1]?.saldo_final +
-              parseInt(req.body.monto),
+        getSaldoEgreso[getSaldoEgreso.length - 1]?.saldo_final === 0
+        ? getSaldoEgreso[getSaldoEgreso.length - 1]?.saldo_inicial +
+          parseInt(req.body.monto)
+        : getSaldoEgreso[getSaldoEgreso.length - 1]?.saldo_final +
+          parseInt(req.body.monto),
       };
       let newSaldoEgreso = {
         saldo_inicial: getSaldo[getSaldo.length - 1]?.saldo_inicial,
@@ -173,9 +173,9 @@ const postIngresoEgreso = async (req, res, next) => {
         egresos: "",
         saldo_final:
           getSaldo[getSaldo.length - 1]?.saldo_final === 0
-            ? getSaldo[getSaldo.length - 1]?.saldo_inicial -
+            ? getSaldo[getSaldo.length - 1]?.saldo_inicial +
               parseInt(req.body.monto)
-            : getSaldo[getSaldo.length - 1]?.saldo_final -
+            : getSaldo[getSaldo.length - 1]?.saldo_final +
               parseInt(req.body.monto),
       };
       const post = await ingresos_egresos.create(info);
@@ -257,20 +257,18 @@ const updateIngresoEgreso = async (req, res, next) => {
     });
 
     if (req.body.movimiento === "Ingreso") {
+      let saldoInicial = parseInt(getSaldo[getSaldo.length - 1]?.saldo_inicial);
+      let saldoFinal = parseInt(getSaldo[getSaldo.length - 1]?.saldo_final);
+      let ingresoActual = parseInt(getSaldo[getSaldo.length - 1]?.ingresos);
+      let montoAnterior = parseInt(getIngresos[getIngresos.length - 1]?.monto);
+      let montoNuevo = parseInt(req.body.monto);
       let newSaldoIngreso = {
-        saldo_inicial: getSaldo[getSaldo.length - 1]?.saldo_inicial,
-        ingresos:
-          getSaldo[getSaldo.length - 1]?.ingresos -
-          getIngresos[getIngresos.length - 1]?.monto +
-          parseInt(req.body.monto),
+        saldo_inicial: saldoInicial,
+        ingresos: ingresoActual - montoAnterior + montoNuevo,
         saldo_final:
           getSaldo[getSaldo.length - 1]?.saldo_final === 0
-            ? getSaldo[getSaldo.length - 1]?.saldo_inicial -
-              getIngresos[getIngresos.length - 1]?.monto +
-              parseInt(req.body.monto)
-            : getSaldo[getSaldo.length - 1]?.saldo_final -
-              getIngresos[getIngresos.length - 1]?.monto +
-              parseInt(req.body.monto),
+            ? saldoInicial - montoAnterior + montoNuevo
+            : saldoFinal - montoAnterior + montoNuevo,
       };
 
       let update = await ingresos_egresos.update(req.body, {
@@ -283,23 +281,23 @@ const updateIngresoEgreso = async (req, res, next) => {
         .status(200)
         .json({ msg: "Movimiento actualizado con éxito!", status: 200 });
       next();
-    } else if (req.body.movimiento === "Egreso") {
+    }
+    if (req.body.movimiento === "Egreso") {
       let newSaldoEgreso = {
         saldo_inicial: getSaldo[getSaldo.length - 1]?.saldo_inicial,
         egresos:
-          getSaldo[getSaldo.length - 1]?.ingresos +
-          getIngresos[getIngresos.length - 1]?.monto +
+          parseInt(getSaldo[getSaldo.length - 1]?.egresos) -
+          parseInt(getIngresos[getIngresos.length - 1]?.egresos) +
           parseInt(req.body.monto),
         saldo_final:
           getSaldo[getSaldo.length - 1]?.saldo_final === 0
             ? getSaldo[getSaldo.length - 1]?.saldo_inicial -
               getIngresos[getIngresos.length - 1]?.monto +
               parseInt(req.body.monto)
-            : getSaldo[getSaldo.length - 1]?.saldo_final +
-              getIngresos[getIngresos.length - 1]?.monto +
+            : parseInt(getSaldo[getSaldo.length - 1]?.saldo_final) +
+              parseInt(getIngresos[getIngresos.length - 1]?.monto) -
               parseInt(req.body.monto),
       };
-
       let update = await ingresos_egresos.update(req.body, {
         where: { id: req.body.id },
       });
@@ -320,17 +318,59 @@ const updateIngresoEgreso = async (req, res, next) => {
 const deleteIngresoEgreso = async (req, res, next) => {
   let id = req.params.id;
   try {
-    // let getIngresos = await ingresos_egresos.findAll({
-    //   where: { id: id },
-    // });
-    let destroy = await ingresos_egresos.destroy({ where: { id: id } });
-    // console.log(getIngresos);
-    res.status(200).json({
-      msg: "Movimiento eliminado con éxito!",
-      status: 200,
-      data: getIngresos,
+    let getIngresos = await ingresos_egresos.findAll({
+      where: { id: id },
     });
-    next();
+
+    let getSaldo = await saldo.findAll({
+      where: { sucursal_id: getIngresos[getIngresos.length - 1]?.sucursal_id },
+    });
+
+    let movimiento = getIngresos[getIngresos.length - 1]?.movimiento;
+    if (movimiento === "Ingreso") {
+      newSaldoIngreso = {
+        ingresos:
+          parseInt(getSaldo[getSaldo.length - 1]?.ingresos) -
+          parseInt(getSaldo[getSaldo.length - 1]?.ingresos),
+        saldo_final:
+          parseInt(getSaldo[getSaldo.length - 1]?.saldo_final) -
+          parseInt(getSaldo[getSaldo.length - 1]?.ingresos),
+      };
+      let destroy = await ingresos_egresos.destroy({ where: { id: id } });
+      const updateSaldo = await saldo.update(newSaldoIngreso, {
+        where: {
+          sucursal_id: getIngresos[getIngresos.length - 1]?.sucursal_id,
+        },
+      });
+      res.status(200).json({
+        msg: "Movimiento eliminado con éxito!",
+        status: 200,
+      });
+      next();
+    }
+
+    if (movimiento === "Egreso") {
+      newSaldoEgreso = {
+        egresos:
+          parseInt(getSaldo[getSaldo.length - 1]?.egresos) -
+          parseInt(getSaldo[getSaldo.length - 1]?.egresos),
+        saldo_final:
+          parseInt(getSaldo[getSaldo.length - 1]?.saldo_final) +
+          parseInt(getSaldo[getSaldo.length - 1]?.egresos),
+      };
+
+      let destroy = await ingresos_egresos.destroy({ where: { id: id } });
+      const updateSaldo = await saldo.update(newSaldoEgreso, {
+        where: {
+          sucursal_id: getIngresos[getIngresos.length - 1]?.sucursal_id,
+        },
+      });
+      res.status(200).json({
+        msg: "Movimiento eliminado con éxito!",
+        status: 200,
+      });
+      next();
+    }
   } catch (error) {
     res.status(500).json({ msg: "No se pudo eliminar.", status: 500 });
   }
@@ -340,12 +380,22 @@ const reporteIngreso = async (req, res, next) => {
   let id = req.params.id;
 
   try {
-    const getIngresoEgresos = await ingresos_egresos.findAll({
-      where: {
+    let filter;
+    if (!req.body.area || req.body.area === "-1") {
+      filter = {
+        sucursal_id: id,
+        fecha: { [Op.between]: [req.body.fecha_inicio, req.body.fecha_fin] },
+      };
+    } else {
+      filter = {
         sucursal_id: id,
         area: req.body.area,
         fecha: { [Op.between]: [req.body.fecha_inicio, req.body.fecha_fin] },
-      },
+      };
+    }
+
+    const getIngresoEgresos = await ingresos_egresos.findAll({
+      where: filter,
     });
 
     const newObj = getIngresoEgresos.reduce(function (acc, currentValue) {
@@ -364,12 +414,9 @@ const reporteIngreso = async (req, res, next) => {
           .map((data) => data)
           .reduce((acc, value) => {
             const item = acc.find((it) => it.movimiento === value.movimiento);
-            console.log(item?.monto);
             item
               ? (item.monto = parseInt(item.monto) + parseInt(value.monto))
-              : // : item=== undefined
-                // ? item.monto =0
-                acc.push(value);
+              : acc.push(value);
             return acc;
           }, [])
       )
@@ -379,12 +426,22 @@ const reporteIngreso = async (req, res, next) => {
 
     let labels = getLabels;
     // terminar falta agregar 0 a los ingresos y egresos cuando no hay nada en el dia
+
     let ingresos = {
       label: "Ingresos",
       type: "line",
-      data: final3
-        .filter((item) => item.movimiento === "Ingreso")
-        .map((item) => item.monto),
+      data: getLabels.map((item, i) =>
+        final3
+          .filter((data, index) => data.movimiento === "Ingreso")
+          .filter((dat) => dat.fecha === item).length === 0
+          ? 0
+          : parseInt(
+              final3
+                .filter((data, index) => data.movimiento === "Ingreso")
+                .filter((dat) => dat.fecha === item)
+                .map((da) => da.monto)
+            )
+      ),
       fill: false,
       borderColor: "rgb(75, 110, 185)",
       tension: 0.1,
@@ -393,27 +450,24 @@ const reporteIngreso = async (req, res, next) => {
     let egresos = {
       label: "Egresos",
       type: "line",
-      data:
-        // getLabels.map((item,i) => {
-        //   return {
-        //     fecha: item,
-        //     monto: final3
-        //       .filter((data, index) => data.movimiento === "Egresos")
-        //       .map((dat, index) => dat.monto)[i]
-
-        //   }
-        // })
+      data: getLabels.map((item, i) =>
         final3
-          .filter((item, i) => item.movimiento === "Egreso")
-          .map((item, i) => item.monto),
-
+          .filter((data, index) => data.movimiento === "Egreso")
+          .filter((dat) => dat.fecha === item).length === 0
+          ? 0
+          : parseInt(
+              final3
+                .filter((data, index) => data.movimiento === "Egreso")
+                .filter((dat) => dat.fecha === item)
+                .map((da) => da.monto)
+            )
+      ),
       fill: false,
       borderColor: "rgb(222, 101, 92)",
       tension: 0.1,
     };
 
     const concat = { labels: labels, ingresos: ingresos, egresos: egresos };
-
     res.status(200).json({ data: concat });
     next();
   } catch (error) {
@@ -423,67 +477,223 @@ const reporteIngreso = async (req, res, next) => {
 
 const convertJsonToExcel = async (req, res, next) => {
   let id = req.params.id;
-
   try {
     const getIngresos = await ingresos_egresos.findAll({
       where: {
         sucursal_id: id,
         fecha: {
-          [Op.between]: [req.params.fecha_inicio, req.params.fecha_fin],
+          [Op.between]: [req.query.fecha_inicio, req.query.fecha_fin],
         },
       },
     });
-    const excelData = getIngresos.map((item) => {
-      return [
-        item.fecha,
-        item.comprobante,
-        item.proveedor,
-        item.descripcion,
-        item.sucursal_id,
-        item.area,
-        item.movimiento,
-        item.encargado,
-        item.monto,
-        item.monto,
-      ];
-    });
+    console.log(getIngresos);
+    const consejoAdministracion = getIngresos
+      .filter((item) => item.area === "Consejo de Administración")
+      .map((item) => {
+        return [
+          item.fecha,
+          item.comprobante,
+          item.proveedor,
+          item.descripcion,
+          item.area,
+          item.encargado,
+          item.movimiento,
+          item.area,
+          item.ingresos ? item.monto : "",
+          item.egresos ? item.monto : "",
+          item.saldo_final,
+        ];
+      });
+
+    const administracionMina = getIngresos
+      .filter((item) => item.area === "Administración Mina ")
+      .map((item) => {
+        return [
+          item.fecha,
+          item.comprobante,
+          item.proveedor,
+          item.descripcion,
+          item.area,
+          item.encargado,
+          item.movimiento,
+          item.area,
+          item.ingresos ? item.monto : "",
+          item.egresos ? item.monto : "",
+          item.saldo_final,
+        ];
+      });
+
+    const operacionMina = getIngresos
+      .filter((item) => item.area === "Operaciones Mina")
+      .map((item) => {
+        return [
+          item.fecha,
+          item.comprobante,
+          item.proveedor,
+          item.descripcion,
+          item.area,
+          item.encargado,
+          item.movimiento,
+          item.area,
+          item.ingresos ? item.monto : "",
+          item.egresos ? item.monto : "",
+          item.saldo_final,
+        ];
+      });
+    const despachos = getIngresos
+      .filter((item) => item.area === "Despachos")
+      .map((item) => {
+        return [
+          item.fecha,
+          item.comprobante,
+          item.proveedor,
+          item.descripcion,
+          item.area,
+          item.encargado,
+          item.movimiento,
+          item.area,
+          item.ingresos ? item.monto : "",
+          item.egresos ? item.monto : "",
+          item.saldo_final,
+        ];
+      });
+
+    const planeamiento = getIngresos
+      .filter((item) => item.area === "Planeamiento")
+      .map((item) => {
+        return [
+          item.fecha,
+          item.comprobante,
+          item.proveedor,
+          item.descripcion,
+          item.area,
+          item.encargado,
+          item.movimiento,
+          item.area,
+          item.ingresos ? item.monto : "",
+          item.egresos ? item.monto : "",
+          item.saldo_final,
+        ];
+      });
+
+    const planta = getIngresos
+      .filter((item) => item.area === "Planta")
+      .map((item) => {
+        return [
+          item.fecha,
+          item.comprobante,
+          item.proveedor,
+          item.descripcion,
+          item.area,
+          item.encargado,
+          item.movimiento,
+          item.area,
+          item.ingresos ? item.monto : "",
+          item.egresos ? item.monto : "",
+          item.saldo_final,
+        ];
+      });
+
+    const sso = getIngresos
+      .filter((item) => item.area === "SSO")
+      .map((item) => {
+        return [
+          item.fecha,
+          item.comprobante,
+          item.proveedor,
+          item.descripcion,
+          item.area,
+          item.encargado,
+          item.movimiento,
+          item.area,
+          item.ingresos ? item.monto : "",
+          item.egresos ? item.monto : "",
+          item.saldo_final,
+        ];
+      });
+
+    const medioAmbiente = getIngresos
+      .filter((item) => item.area === "Medio Ambiente")
+      .map((item) => {
+        return [
+          item.fecha,
+          item.comprobante,
+          item.proveedor,
+          item.descripcion,
+          item.area,
+          item.encargado,
+          item.movimiento,
+          item.area,
+          item.ingresos ? item.monto : "",
+          item.egresos ? item.monto : "",
+          item.saldo_final,
+        ];
+      });
 
     const workSheetColumnsName = [
-      "Fecha",
-      "Comprobante",
-      "Proveedor",
-      "Concepto",
-      "Caja",
-      "Área",
-      "Tipo de Gasto",
-      "Resp. de Gasto",
-      "Ingreso",
-      "Egreso",
-      "Saldo",
+      "FECHA",
+      "COMPROBANTE",
+      "PROVEEDOR",
+      "CONCEPTO",
+      "CAJA",
+      "ÁREA",
+      "TIPO DE GASTO",
+      "RESP. DE GASTO",
+      "INGRESO",
+      "EGRESO",
+      "SALDO",
     ];
 
     const workBook = XLSX.utils.book_new();
-    const workSheetData = [workSheetColumnsName, ...excelData];
 
-    const workSheet = XLSX.utils.aoa_to_sheet(workSheetData);
+    //Consejo administracion
+    const workSheetData1 = [workSheetColumnsName, ...consejoAdministracion];
+    const workSheet1 = XLSX.utils.aoa_to_sheet(workSheetData1);
     XLSX.utils.book_append_sheet(
       workBook,
-      workSheet,
+      workSheet1,
       "Consejo de Administración"
     );
-    XLSX.utils.book_append_sheet(workBook, workSheet, "Administración mina");
-    XLSX.utils.book_append_sheet(workBook, workSheet, "Operaciones mina");
-    XLSX.utils.book_append_sheet(workBook, workSheet, "Despachos");
-    XLSX.utils.book_append_sheet(workBook, workSheet, "Planeamiento");
-    XLSX.utils.book_append_sheet(workBook, workSheet, "Planta");
 
-    XLSX.utils.book_append_sheet(workBook, workSheet, "SSO");
-    XLSX.utils.book_append_sheet(workBook, workSheet, "Medio Ambiente");
+    //Administracion mina
+    const workSheetData2 = [workSheetColumnsName, ...administracionMina];
+    const workSheet2 = XLSX.utils.aoa_to_sheet(workSheetData2);
+    XLSX.utils.book_append_sheet(workBook, workSheet2, "Administración mina");
 
-    const filePath = path.resolve("./reporte.xlsx");
-    const file = new Blob([filePath], { type: "application/vnd.ms-excel" });
-    console.log(file);
-    res.status(200).download(filePath);
+    //Operaciones mina
+    const workSheetData3 = [workSheetColumnsName, ...operacionMina];
+    const workSheet3 = XLSX.utils.aoa_to_sheet(workSheetData3);
+    XLSX.utils.book_append_sheet(workBook, workSheet3, "Operaciones mina");
+
+    //Despachos
+    const workSheetData4 = [workSheetColumnsName, ...despachos];
+    const workSheet4 = XLSX.utils.aoa_to_sheet(workSheetData4);
+    XLSX.utils.book_append_sheet(workBook, workSheet4, "Despachos");
+
+    //Planeamiento
+    const workSheetData5 = [workSheetColumnsName, ...planeamiento];
+    const workSheet5 = XLSX.utils.aoa_to_sheet(workSheetData5);
+    XLSX.utils.book_append_sheet(workBook, workSheet5, "Planeamiento");
+
+    //Planta
+    const workSheetData6 = [workSheetColumnsName, ...planta];
+    const workSheet6 = XLSX.utils.aoa_to_sheet(workSheetData6);
+    XLSX.utils.book_append_sheet(workBook, workSheet6, "Planta");
+
+    //SSO
+    const workSheetData7 = [workSheetColumnsName, ...sso];
+    const workSheet7 = XLSX.utils.aoa_to_sheet(workSheetData7);
+    XLSX.utils.book_append_sheet(workBook, workSheet7, "SSO");
+
+    //SSO
+    const workSheetData8 = [workSheetColumnsName, ...medioAmbiente];
+    const workSheet8 = XLSX.utils.aoa_to_sheet(workSheetData8);
+    XLSX.utils.book_append_sheet(workBook, workSheet8, "Medio Ambiente");
+
+    XLSX.write(workBook, { bookType: "xlsx", type: "binary" });
+    XLSX.writeFile(workBook, "reporte.xlsx");
+
+    res.status(200).sendFile("reporte.xlsx", { root: "." });
   } catch (error) {
     console.log(error);
     res.status(500).json(error);
